# src/fleet/ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/tasks": {
      "get": {
        "summary": "List Tasks",
        "description": "List task summaries, active first then recently-closed (FR-07).",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "closed_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "title": "Closed Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "description": "Create a task via the queue; 201 with the new id.",
        "operationId": "create_task_api_tasks_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/beads/{bead_id}": {
      "get": {
        "summary": "Get Bead",
        "description": "One bead with description, notes, dependencies, comments.",
        "operationId": "get_bead_api_beads__bead_id__get",
        "parameters": [
          {
            "name": "bead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/beads/{bead_id}/status": {
      "post": {
        "summary": "Set Status",
        "description": "Set a bead status; `closed` goes through `bd close` like queue.close.",
        "operationId": "set_status_api_beads__bead_id__status_post",
        "parameters": [
          {
            "name": "bead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/beads/{bead_id}/unblock": {
      "post": {
        "summary": "Unblock Bead",
        "description": "Reopen a bead (`bd update --status open`).",
        "operationId": "unblock_bead_api_beads__bead_id__unblock_post",
        "parameters": [
          {
            "name": "bead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/beads/{bead_id}/remove-assignee": {
      "post": {
        "summary": "Remove Bead Assignee",
        "description": "Clear the assignee; the portal targets beads, so a missing task dir is ok.",
        "operationId": "remove_bead_assignee_api_beads__bead_id__remove_assignee_post",
        "parameters": [
          {
            "name": "bead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeadComment": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "text",
          "created_at"
        ],
        "title": "BeadComment",
        "description": "One comment on a bead detail."
      },
      "BeadDependency": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "dependency_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "dependency_type"
        ],
        "title": "BeadDependency",
        "description": "One dependency edge on a bead detail."
      },
      "BeadDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "issue_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Type"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/BeadDependency"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/BeadComment"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "issue_type",
          "assignee",
          "description",
          "notes",
          "created_at",
          "updated_at",
          "closed_at",
          "close_reason",
          "dependencies",
          "comments"
        ],
        "title": "BeadDetail",
        "description": "One bead with description, notes, dependencies and comments."
      },
...
      "CreateTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateTaskResponse",
        "description": "New task id for POST /api/tasks."
      },
...
      "FollowupItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          }
        },
        "type": "object",
        "title": "FollowupItem",
        "description": "One observer-declared follow-up bead inside a partial result."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobArtifacts": {
        "properties": {
          "research": {
            "type": "boolean",
            "title": "Research"
          },
          "design": {
            "type": "boolean",
            "title": "Design"
          },
          "tasks": {
            "type": "boolean",
            "title": "Tasks"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "research",
          "design",
          "tasks",
          "approved"
        ],
        "title": "JobArtifacts",
        "description": "Presence flags for the job worker's RESEARCH/DESIGN/tasks/APPROVED."
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "description": "Generic mutation acknowledgement ({\"ok\": true})."
      },
...
      "RunStep": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "fail",
              "outcome"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name",
          "started_at",
          "ended_at",
          "status",
          "reason"
        ],
        "title": "RunStep",
        "description": "One entry of the current attempt's run.json steps."
      },
...
      "TaskDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "coder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coder"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "elapsed_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Sec"
          },
          "idle_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Sec"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Tokens"
          },
          "context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Pct"
          },
          "context_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Limit"
          },
          "last_event_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Kind"
          },
          "last_event_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Detail"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "blocked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked At"
          },
          "ignore_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ignore Until"
          },
          "ignored": {
            "type": "boolean",
            "title": "Ignored"
          },
          "rounds": {
            "$ref": "#/components/schemas/TaskRounds"
          },
          "restarts": {
            "type": "integer",
            "title": "Restarts"
          },
          "context_rounds": {
            "type": "integer",
            "title": "Context Rounds"
          },
          "compactions": {
            "type": "integer",
            "title": "Compactions"
          },
          "peak_context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Context Pct"
          },
          "last_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome"
          },
          "last_outcome_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome Reason"
          },
          "last_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Excerpt"
          },
          "worker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker"
          },
          "job_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Phase"
          },
          "job_artifacts": {
            "$ref": "#/components/schemas/JobArtifacts"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RunStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "lease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskLease"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/TaskAttempt"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "cwd",
          "coder",
          "model",
          "priority",
          "depends_on",
          "created_at",
          "started_at",
          "ended_at",
          "elapsed_sec",
          "idle_sec",
          "events",
          "context_tokens",
          "context_pct",
          "context_limit",
          "last_event_kind",
          "last_event_detail",
          "blocked_reason",
          "blocked_at",
          "ignore_until",
          "ignored",
          "rounds",
          "restarts",
          "context_rounds",
          "compactions",
          "peak_context_pct",
          "last_outcome",
          "last_outcome_reason",
          "last_action",
          "result",
          "state_excerpt",
          "worker",
          "job_phase",
          "job_artifacts",
          "steps",
          "lease",
          "attempts"
        ],
        "title": "TaskDetail",
        "description": "One task with its attempts timeline (GET /api/tasks/{id})."
      },
...
      "TaskLease": {
        "properties": {
          "heartbeat_at": {
            "type": "string",
            "title": "Heartbeat At"
          },
          "lease_until": {
            "type": "string",
            "title": "Lease Until"
          },
          "alive": {
            "type": "boolean",
            "title": "Alive"
          }
        },
        "type": "object",
        "required": [
          "heartbeat_at",
          "lease_until",
          "alive"
        ],
        "title": "TaskLease",
        "description": "Claim lease from the latest attempt's run.json, null when absent."
      },
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskDetail"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "Envelope for GET /api/tasks."
      },
      "TaskResult": {
        "properties": {
          "schema": {
            "type": "integer",
            "title": "Schema"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "partial",
              "blocked"
            ],
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "commits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commits"
          },
          "tests": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          },
          "blocked_reason": {
            "type": "string",
            "title": "Blocked Reason"
          },
          "followups": {
            "items": {
              "$ref": "#/components/schemas/FollowupItem"
            },
            "type": "array",
            "title": "Followups"
          }
        },
        "type": "object",
        "required": [
          "schema",
          "status",
          "summary",
          "commits",
          "tests",
          "open_questions",
          "next_step",
          "blocked_reason"
        ],
        "title": "TaskResult",
        "description": "The worker's declared outcome (live RESULT.json or latest snapshot)."
      },
      "TaskRounds": {
        "properties": {
          "failure": {
            "type": "integer",
            "title": "Failure"
          },
          "stall": {
            "type": "integer",
            "title": "Stall"
          },
          "context": {
            "type": "integer",
            "title": "Context"
          },
          "partial": {
            "type": "integer",
            "title": "Partial"
          },
          "noclose": {
            "type": "integer",
            "title": "Noclose"
          }
        },
        "type": "object",
        "required": [
          "failure",
          "stall",
          "context",
          "partial",
          "noclose"
        ],
        "title": "TaskRounds",
        "description": "Trailing-streak retry rounds, from attempts.jsonl."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}